+++ src/main.rs
@@ -9,7 +9,7 @@ use std::{
     error::Error,
     fs,
     io::{self, Read},
-    path::PathBuf,
+    path::{Path, PathBuf},
 };
 
 use rlua::{Lua, MultiValue, ToLua};
@@ -42,23 +42,7 @@ fn start() -> Result<(), Box<dyn Error>> {
     env_logger::from_env(env_logger::Env::default().default_filter_or("warn")).init();
 
     let opt = Options::from_args();
-
-    let (contents, chunk_name) = if opt.script.as_os_str() == "-" {
-        let mut contents = String::new();
-        io::stdin().read_to_string(&mut contents)?;
-
-        (contents, "stdin".to_owned())
-    } else {
-        let contents = fs::read_to_string(&opt.script)?;
-        let file_name = opt
-            .script
-            .file_name()
-            .unwrap()
-            .to_string_lossy()
-            .into_owned();
-
-        (contents, file_name)
-    };
+    let (contents, chunk_name) = load_script(&opt.script)?;
 
     let lua = Lua::new();
 
@@ -82,6 +66,63 @@ fn start() -> Result<(), Box<dyn Error>> {
     })?;
 
     Ok(())
+}
+
+/// Load the script from the given CLI-supplied path.
+///
+/// Returns the contents of the script followed by its chunk name that should be
+/// given to Lua.
+fn load_script(script: &Path) -> io::Result<(String, String)> {
+    // Passing `-` indicates that the script should be read from stdin.
+    if script.as_os_str() == "-" {
+        let mut contents = String::new();
+        io::stdin().read_to_string(&mut contents)?;
+
+        return Ok((contents, "stdin".to_owned()));
+    }
+
+    log::trace!("Reading script from {}", script.display());
+
+    match fs::read_to_string(script) {
+        // If the input is an exact file name that exists, we'll run that
+        // script.
+        Ok(contents) => {
+            let file_name = script.file_name().unwrap().to_string_lossy().into_owned();
+
+            Ok((contents, file_name))
+        }
+
+        Err(err) => {
+            // If we couldn't find the exact name we're looking for, but a
+            // script with that name inside a folder named `.remodel/` exists,
+            // we'll run that!
+            if script.is_relative() {
+                if let Some(command_name) = script.file_name().and_then(|name| name.to_str()) {
+                    let mut command_path = PathBuf::from(".remodel");
+                    command_path.push(format!("{}.lua", command_name));
+
+                    log::trace!("Reading script from {}", command_path.display());
+
+                    match fs::read_to_string(command_path) {
+                        Ok(contents) => {
+                            return Ok((contents, command_name.to_owned()));
+                        }
+
+                        // If we got any IO errors that weren't 'file not
+                        // found', we'll report those. Otherwise, reporting an
+                        // error with this file path might confuse the user.
+                        Err(err) => {
+                            if err.kind() != io::ErrorKind::NotFound {
+                                return Err(err);
+                            }
+                        }
+                    }
+                }
+            }
+
+            Err(err)
+        }
+    }
 }
 
 fn main() {